requirements.txt
# requirements.txt
boto3>=1.26.0
pyyaml>=5.1
colorama>=0.4.4
kubernetes>=12.0.0
cryptography>=3.4.7
